# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Addresses list",
                "operationId": "a9af9820257150bfb712d6851649fe72",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Count of addresses per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AddressResource"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Create address",
                "operationId": "f10f32387b2ddc7915e53ecc6ccb7057",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "country",
                                    "city",
                                    "street",
                                    "house"
                                ],
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Russia"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Tver oblast"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Kalyazin"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Kominterna"
                                    },
                                    "house": {
                                        "type": "string",
                                        "example": "105/2"
                                    },
                                    "flat": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Address"
                ],
                "summary": "Update address",
                "operationId": "76899dc00e38f1bdf3e1b605987b0cc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string",
                                        "example": "Russia"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "Tver oblast"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Kalyazin"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Kominterna"
                                    },
                                    "house": {
                                        "type": "string",
                                        "example": "105/2"
                                    },
                                    "flat": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/addresses/{id}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get address by ID",
                "operationId": "7c9c5679531c1f3dea1805a3edc7a0c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "Russia"
                                        },
                                        "subject": {
                                            "type": "string",
                                            "example": "Tver oblast"
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Kalyazin"
                                        },
                                        "street": {
                                            "type": "string",
                                            "example": "Kominterna"
                                        },
                                        "house": {
                                            "type": "string",
                                            "example": "105/2"
                                        },
                                        "flat": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2020-01-27T17:50:45Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2020-01-27T17:50:45Z"
                                        },
                                        "link": {
                                            "type": "string",
                                            "example": "http://localhost:8080/api/v1/addresses/1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Pagination": {
                "properties": {
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 4
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://localhost:8080/api/v1/****?page=2"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
            "AddressResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "country": {
                        "type": "string",
                        "example": "Russia"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Tver oblast"
                    },
                    "city": {
                        "type": "string",
                        "example": "Kalyazin"
                    },
                    "street": {
                        "type": "string",
                        "example": "Kominterna"
                    },
                    "house": {
                        "type": "string",
                        "example": "105/2"
                    },
                    "flat": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-27T17:50:45Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-27T17:50:45Z"
                    },
                    "link": {
                        "type": "string",
                        "readOnly": true,
                        "example": "http://localhost:8080/api/v1/addresses/1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token as: Bearer {your-token}"
            }
        }
    },
...
}